+++ zarrs/src/array/array_ops/async_array_read_ops/array.rs
@@ -16,7 +16,7 @@ use super::AsyncArrayReadOps;
 use crate::array::{ArrayBytes, ArraySubset, ChunkShapeTraits};
 use zarrs_codec::{
     ArrayBytesDecodeIntoTarget, ArrayToBytesCodecTraits, AsyncArrayPartialDecoderTraits,
-    AsyncStoragePartialDecoder, CodecError, InvalidNumberOfElementsError, copy_fill_value_into,
+    CodecError, InvalidNumberOfElementsError, copy_fill_value_into,
 };
 use zarrs_storage::{Bytes, StorageHandle};
 
@@ -109,10 +109,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> AsyncArrayReadOps
                 .storage_transformers()
                 .create_async_readable_transformer(storage_handle)
                 .await?;
-            let input_handle = Arc::new(AsyncStoragePartialDecoder::new(
-                storage_transformer,
-                self.chunk_key(chunk_indices),
-            ));
+            let input_handle = Arc::new((storage_transformer, self.chunk_key(chunk_indices)));
             let bytes = self
                 .codecs_bound()
                 .async_partial_decoder(input_handle, &chunk_shape, options)
@@ -440,10 +437,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> AsyncArrayReadOps
             .storage_transformers()
             .create_async_readable_transformer(storage_handle)
             .await?;
-        let input_handle = Arc::new(AsyncStoragePartialDecoder::new(
-            storage_transformer,
-            self.chunk_key(chunk_indices),
-        ));
+        let input_handle = Arc::new((storage_transformer, self.chunk_key(chunk_indices)));
         Ok(self
             .codecs_bound()
             .async_partial_decoder(input_handle, &self.chunk_shape(chunk_indices)?, options)
@@ -753,11 +747,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> Array<TStorage> {
                 .storage_transformers()
                 .create_async_readable_transformer(storage_handle)
                 .await?;
-            let input_handle = Arc::new(AsyncStoragePartialDecoder::new(
-                storage_transformer,
-                self.chunk_key(chunk_indices),
-            ));
-
+            let input_handle = Arc::new((storage_transformer, self.chunk_key(chunk_indices)));
             self.codecs_bound()
                 .async_partial_decoder(input_handle, &chunk_shape, options)
                 .await?