+++ zarrs/src/array/codec/array_to_array/squeeze/squeeze_codec_partial.rs
@@ -75,10 +75,6 @@ impl<T: ?Sized> ArrayPartialEncoderTraits for SqueezeCodecPartial<T>
 where
     T: ArrayPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn ArrayPartialDecoderTraits> {
-        self.clone()
-    }
-
     fn erase(&self) -> Result<(), CodecError> {
         self.input_output_handle.erase()
     }
@@ -154,10 +150,6 @@ impl<T: ?Sized> AsyncArrayPartialEncoderTraits for SqueezeCodecPartial<T>
 where
     T: AsyncArrayPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn AsyncArrayPartialDecoderTraits> {
-        self.clone()
-    }
-
     async fn erase(&self) -> Result<(), CodecError> {
         self.input_output_handle.erase().await
     }