+++ zarrs/src/array/codec/array_to_array/transpose/transpose_codec_partial.rs
@@ -140,10 +140,6 @@ impl<T: ?Sized> ArrayPartialEncoderTraits for TransposeCodecPartial<T>
 where
     T: ArrayPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn ArrayPartialDecoderTraits> {
-        self.clone()
-    }
-
     fn erase(&self) -> Result<(), CodecError> {
         self.input_output_handle.erase()
     }
@@ -259,10 +255,6 @@ impl<T: ?Sized> AsyncArrayPartialEncoderTraits for TransposeCodecPartial<T>
 where
     T: AsyncArrayPartialEncoderTraits,
 {
-    fn into_dyn_decoder(self: Arc<Self>) -> Arc<dyn AsyncArrayPartialDecoderTraits> {
-        self.clone()
-    }
-
     async fn erase(&self) -> Result<(), CodecError> {
         self.input_output_handle.erase().await
     }